pyproject.toml
[project]
name = "doc2md"
version = "0.1.0"
description = "Convert generic documents (including docx and pdf with images and math formulas) into markdown folders."
readme = "README.md"
authors = [
    { name = "Matteo Bortolotto", email = "mbortolotto@fbk.eu" }
]
requires-python = ">=3.12"
dependencies = [
    "docling>=2.75.0",
    "torchvision>=0.25.0",
    "click>=8.0",
    "reportlab>=4.4.10",
]

[project.scripts]
doc2md = "doc2md.cli:main"

[project.optional-dependencies]
dev = [
    "pytest",
    "python-docx",
    "Pillow",
    "reportlab",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
